Refactor data models
Move to .core.
