--- src/ccompass/PPMS.py
@@ -1,37 +1,0 @@
-"""Preparameters"""
-
-
-def fract_default():
-    params_default = {
-        "class": {
-            "scale1": [
-                True,
-                "area",
-            ],
-            "corrfilter": False,
-            "scale2": [False, "area"],
-            "zeros": True,
-            "combination": "separate",
-        },
-        "vis": {
-            "scale1": [
-                True,
-                "minmax",
-            ],
-            "corrfilter": False,
-            "scale2": [True, "minmax"],
-            "zeros": True,
-            "combination": "median",
-        },
-        "global": {
-            "missing": [True, "1"],
-            "minrep": [True, "2"],
-            "outcorr": False,
-        },
-    }
-    return params_default
-
-
-def tp_default():
-    params_default = {"minrep": 2, "imputation": "normal"}
-    return params_default
